% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PathwayAnalysis.R
\name{Base2IndVec}
\alias{Base2IndVec}
\title{Base2IndVec}
\usage{
Base2IndVec(vec)
}
\arguments{
\item{vec}{a binary genotype vector}
}
\value{
a number used for indexing a given genotype
}
\description{
Base2IndVec
}
\examples{
vec<-c(0,1,0,1)
Base2IndVec(vec)
}
