% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcbn.R
\name{bcbn}
\alias{bcbn}
\title{B-CBN}
\usage{
bcbn(
  data = NULL,
  n_samples = 25000,
  theta = 0,
  epsilon = 0.05,
  n_chains = 4,
  thin = 10,
  Max_L = 1000,
  n_cores = 1
)
}
\arguments{
\item{data}{Generated data}

\item{n_samples}{Number of samples <def: 25000>}

\item{theta}{Theta <def: 0>}

\item{epsilon}{Epsilon <def: 0.05>}

\item{n_chains}{N-Chains <def: 4>}

\item{thin}{Thin <def: 10>}

\item{n_cores}{Number of parallelized cores <def: 1>}
}
\value{
A matrix
}
\description{
B-CBN
}
\examples{
if (!require("rBCBN")) {
    install.packages(CBN2Path::getBCBNinstall(), repos = NULL, type = "source")
}
bcbn()
}
