% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PathwayAnalysis.R
\name{combinations}
\alias{combinations}
\title{combinations}
\usage{
combinations(n, r, v = 1:n, set = TRUE, repeats.allowed = FALSE)
}
\arguments{
\item{n}{total number of elements in the set}

\item{r}{subset size}

\item{repeats.allowed}{}
}
\description{
combinations
}
