% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctcbn.R
\name{ctcbn_single}
\alias{ctcbn_single}
\title{CT-CBN Single Batch}
\usage{
ctcbn_single(
  dataset,
  bootstrap_mode = FALSE,
  bootstrap_samples = 0,
  random_seed = 0,
  sampling_rate = 1,
  epsilon = 2,
  num_drawn_samples = 0,
  num_em_runs = 1,
  output_stem = tempfile("output")
)
}
\arguments{
\item{dataset}{\code{Spock} object with poset and pattern/lambda data.}

\item{bootstrap_mode}{Boolean representing bootstrapping mode.}

\item{bootstrap_samples}{Number of bootstrap samples (requires \code{epsilon} > 0, \code{num_drawn_samples} = 0)}

\item{random_seed}{Random seed.}

\item{sampling_rate}{Sampling rate.}

\item{epsilon}{If between 0 and 1, the fraction of violations allowed per edge. If negative, the interval 0 to 0.5 will be sampled equidistantly with N points.}

\item{num_drawn_samples}{If > 0, the number of samples to draw from the model. If zero (default), the model will be learned from data.}

\item{num_em_runs}{Number of em runs.}

\item{output_stem}{File to output to (defaults to a tempfile).}
}
\value{
A list of output data.
}
\description{
CT-CBN Single Batch
}
\examples{
example_path = get_examples()[1]
bc = Spock$new(
  poset = read_poset(example_path)$sets,
  numMutations = read_poset(example_path)$mutations,
  patternOrLambda = read_pattern(example_path)
)
ctcbn_single(bc)
}
