% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PathwayAnalysis.R
\name{generate_matrix_genotypes}
\alias{generate_matrix_genotypes}
\title{generate_matrix_genotypes: Generates a binary matrix of a given length}
\usage{
generate_matrix_genotypes(g)
}
\arguments{
\item{g}{genotype length}
}
\value{
a genotype matrix with ncol=g and nrow=2^g
}
\description{
generate_matrix_genotypes: Generates a binary matrix of a given length
}
