% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PathwayAnalysis.R
\name{generateMatrixGenotypes}
\alias{generateMatrixGenotypes}
\title{generateMatrixGenotypes}
\usage{
generateMatrixGenotypes(g)
}
\arguments{
\item{g}{genotype length}
}
\value{
a genotype matrix with ncol=g and nrow=2^g
}
\description{
generateMatrixGenotypes
}
\examples{
Geno4<-generateMatrixGenotypes(4)
}
