% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PathwayAnalysis.R
\name{genotypeFeasibility}
\alias{genotypeFeasibility}
\title{genotypeFeasibility}
\usage{
genotypeFeasibility(genotypes, dag, x)
}
\arguments{
\item{genotypes}{the full set of potential binary genotypes of a given length.}

\item{dag}{matrix representing the DAG of restrictions.}

\item{x}{the number of mutations considered.}
}
\value{
a binary vector, which indicates feasibility or infeasibility of a set of genotypes
}
\description{
genotypeFeasibility
}
\examples{
geno4 <- generateMatrixGenotypes(4)
dag <- matrix(c(4, 4, 4, 1, 2, 3), 3, 2)
x <- 4
genoF4 <- genotypeFeasibility(geno4, dag, x)
}
