% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PathwayAnalysis.R
\name{genotypeMatrixMutator}
\alias{genotypeMatrixMutator}
\title{genotypeMatrixMutator}
\usage{
genotypeMatrixMutator(mat, fp, fn)
}
\arguments{
\item{mat}{The genotype matrix including sampled genotypes, which need to be mutated.}

\item{fp}{False positive rate}

\item{fn}{False negative rate}
}
\value{
The mutated version of the genotype matrix
}
\description{
genotypeMatrixMutator
}
\examples{
set.seed(100)
gMat <- matrix(sample(c(0, 1), 800, replace = TRUE), 200, 4)
gMatMut <- genotypeMatrixMutator(gMat, 0.2, 0.2)
}
