% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcga.R
\name{getRawTCGAData}
\alias{getRawTCGAData}
\title{Get Raw TCGA Data}
\usage{
getRawTCGAData(project)
}
\arguments{
\item{project}{TCGA project ID; pass "help" to see list of all project IDs}
}
\value{
data frame of TCGA data for given project
}
\description{
Get Raw TCGA Data
}
\examples{
getRawTCGAData("help")
}
