% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hcbn.R
\name{hcbn}
\alias{hcbn}
\title{H-CBN}
\usage{
hcbn(
  datasets,
  anneal = FALSE,
  temp = 0,
  annealing_steps = 0,
  epsilon = 2,
  n_cores = 1
)
}
\arguments{
\item{datasets}{Vector of \code{Spock} objects with poset and pattern/lambda data or a \code{Spock} object (alias of hcbnSingle).}

\item{anneal}{If \code{TRUE}, performes a simulated annealing run starting from the poset}

\item{temp}{Temperature of simulated annealing.}

\item{annealing_steps}{Number of simulated annealing steps.}

\item{epsilon}{Value of eps for CT-CBN model selection. Requires both pattern and lambda data in input \code{Spock}.}

\item{n_cores}{Maximum number of threads to use to parallelize.}
}
\value{
A matrix of results.
}
\description{
H-CBN
}
\examples{
example_path <- getExamples()[1]
bc <- Spock$new(
    poset = readPoset(example_path)$sets,
    numMutations = readPoset(example_path)$mutations,
    genotypeMatrix = readPattern(example_path)
)
hcbn(bc)
hcbn(c(bc, bc, bc))
}
