% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PathwayAnalysis.R
\name{Jensen_Shannon_Divergence}
\alias{Jensen_Shannon_Divergence}
\title{Jensen_Shannon_Divergence}
\usage{
Jensen_Shannon_Divergence(Prob1, Prob2)
}
\arguments{
\item{Prob1}{The first (discrete) probability distribution (vector)}

\item{Prob2}{The second (discrete) probability distribution (vector)}
}
\value{
Jensen Shannon Divergence between the two (discrete) probability distributions
}
\description{
Jensen_Shannon_Divergence
}
