% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PathwayAnalysis.R
\name{jensenShannonDivergence}
\alias{jensenShannonDivergence}
\title{jensenShannonDivergence}
\usage{
jensenShannonDivergence(prob1, prob2)
}
\arguments{
\item{prob1}{The first (discrete) probability distribution (vector)}

\item{prob2}{The second (discrete) probability distribution (vector)}
}
\value{
Jensen Shannon Divergence between the two (discrete) probability distributions
}
\description{
jensenShannonDivergence
}
\examples{
set.seed(100)
gMat <- matrix(sample(c(0, 1), 12, replace = TRUE), 3, 4)
pathCT <- pathProbQuartetCTCBN(gMat)
pathH <- pathProbQuartetHCBN(gMat)
jensenShannonDivergence(pathCT, pathH)
}
