% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PathwayAnalysis.R
\name{pathEdgeMapper}
\alias{pathEdgeMapper}
\title{pathEdgeMapper}
\usage{
pathEdgeMapper(x)
}
\arguments{
\item{x}{number of mutations to consider}
}
\value{
Pathway to edge compatibility matrix, each element of which indicates whether a given edge is included in the transitive closure of a given pathway (1) or not (0).
}
\description{
pathEdgeMapper
}
\examples{
peMap <- pathEdgeMapper(4)
}
