% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PathwayAnalysis.R
\name{pathNormalization}
\alias{pathNormalization}
\title{pathNormalization}
\usage{
pathNormalization(pathProb, x)
}
\arguments{
\item{pathProb}{The pathway probabilities returned in the step 3 of the R-CBN algorithm}

\item{x}{The number of mutations to consider}
}
\value{
The updated pathway probabilities (the step 5 of the R-CBN algorithm)
}
\description{
pathNormalization
}
\examples{
dag <- matrix(c(2, 2, 4, 1, 3, 3), 3, 2)
lambda <- c(1, 4, 3, 2.5, 2)
x <- 4
pathP <- pathProbCBN(dag, lambda, x)
pathN <- pathNormalization(pathP, x)
}
