% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PathwayAnalysis.R
\name{PathProb_CBN}
\alias{PathProb_CBN}
\title{PathProb_CBN: quantifies pathway probabilities using the output of CT-CBN or H-CBN}
\usage{
PathProb_CBN(DAG, LAMBDA, x)
}
\arguments{
\item{DAG}{matrix representing the DAG of restrictions.}

\item{LAMBDA}{the lambda values, which are produced by the CBN model.}

\item{x}{the number of mutations considered.}
}
\value{
vector of probabilities assigned to a set of pathways
}
\description{
PathProb_CBN: quantifies pathway probabilities using the output of CT-CBN or H-CBN
}
