% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PathwayAnalysis.R
\name{PathProb_Quartet_CTCBN}
\alias{PathProb_Quartet_CTCBN}
\title{PathProb_Quartet_CTCBN}
\usage{
PathProb_Quartet_CTCBN(gMat)
}
\arguments{
\item{gMat}{The n by 4 binary genotype matrix representing a given quartet for a sample of n genotypes.}
}
\value{
The probability distribution (returned by the CT-CBN model), which is represented as a vector of length 24.
}
\description{
PathProb_Quartet_CTCBN
}
