% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PathwayAnalysis.R
\name{PathProb_Quartet_HCBN}
\alias{PathProb_Quartet_HCBN}
\title{PathProb_Quartet_HCBN}
\usage{
PathProb_Quartet_HCBN(gMat)
}
\arguments{
\item{gMat}{The n by 4 binary genotype matrix representing a given quartet for a sample of n genotypes.}
}
\value{
The probability distribution (returned by the H-CBN model), which is represented as a vector of length 24.
}
\description{
PathProb_Quartet_HCBN
}
\examples{
set.seed(100)
gMat<-matrix(sample(c(0,1),800,replace = TRUE),200,4)
PathH<-PathProb_Quartet_HCBN(gMat)
}
