% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PathwayAnalysis.R
\name{pathProbQuartetBCBN}
\alias{pathProbQuartetBCBN}
\title{pathProbQuartetBCBN}
\usage{
pathProbQuartetBCBN(gMat)
}
\arguments{
\item{gMat}{The n by 4 binary genotype matrix representing a given quartet for a sample of n genotypes.}
}
\value{
The probability distribution (returned by the B-CBN model), which is represented as a vector of length 24.
}
\description{
pathProbQuartetBCBN
}
\examples{
set.seed(100)
gMat<-matrix(sample(c(0,1),800,replace = TRUE),200,4)
PathB<-pathProbQuartetBCBN(gMat)
}
