% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PathwayAnalysis.R
\name{pathProbQuartetCTCBN}
\alias{pathProbQuartetCTCBN}
\title{pathProbQuartetCTCBN}
\usage{
pathProbQuartetCTCBN(gMat)
}
\arguments{
\item{gMat}{The n by 4 binary genotype matrix representing a given quartet for a sample of n genotypes.}
}
\value{
The probability distribution (returned by the CT-CBN model), which is represented as a vector of length 24.
}
\description{
pathProbQuartetCTCBN
}
\examples{
set.seed(100)
gMat<-matrix(sample(c(0,1),800,replace = TRUE),200,4)
PathCT<-pathProbQuartetCTCBN(gMat)
}
