% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PathwayAnalysis.R
\name{pathProbQuartetHCBN}
\alias{pathProbQuartetHCBN}
\title{pathProbQuartetHCBN}
\usage{
pathProbQuartetHCBN(gMat)
}
\arguments{
\item{gMat}{The n by 4 binary genotype matrix representing a given quartet for a sample of n genotypes.}
}
\value{
The probability distribution (returned by the H-CBN model), which is represented as a vector of length 24.
}
\description{
pathProbQuartetHCBN
}
\examples{
set.seed(100)
gMat<-matrix(sample(c(0,1),800,replace = TRUE),200,4)
PathH<-pathProbQuartetHCBN(gMat)
}
