% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PathwayAnalysis.R
\name{pathProbQuartetRCBN}
\alias{pathProbQuartetRCBN}
\title{pathProbQuartetRCBN}
\usage{
pathProbQuartetRCBN(gMat)
}
\arguments{
\item{gMat}{The n by 4 binary genotype matrix representing a given quartet for a sample of n genotypes.}
}
\value{
The probability distribution (returned by the R-CBN model), which is represented as a vector of length 24
}
\description{
pathProbQuartetRCBN
}
\examples{
set.seed(100)
gMat<-matrix(sample(c(0,1),12,replace = TRUE),3,4)
pathProbQuartetRCBN(gMat)
}
