% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PathwayAnalysis.R
\name{pathProbSSWM}
\alias{pathProbSSWM}
\title{pathProbSSWM}
\usage{
pathProbSSWM(fitness, x)
}
\arguments{
\item{fitness}{A vector of length 2^x, each element of which representing the fitness assigned to one of the 2^x genotypes.}

\item{x}{The number of mutations considered.}
}
\value{
vector of probabilities assigned to all potential pathways of length x
}
\description{
pathProbSSWM
}
\examples{
f <- c(0, 0.1, 0.2, 0.1, 0.2, 0.4, 0.3, 0.2, 0.2, 0.1, 0, 0.6, 0.4, 0.3, 0.2, 1)
x <- 4
pathP <- pathProbSSWM(f, x)
}
