% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PathwayAnalysis.R
\name{pathwayCompatibilityQuartet}
\alias{pathwayCompatibilityQuartet}
\title{pathwayCompatibilityQuartet}
\usage{
pathwayCompatibilityQuartet(gMat)
}
\arguments{
\item{gMat}{The n by 4 binary genotype matrix representing a given quartet for a sample of n genotypes.}
}
\value{
The compatibility score, which is represented as a vector of length 24, each element of which corresponds to one of the 24 pathways of length 4.
}
\description{
pathwayCompatibilityQuartet
}
\examples{
set.seed(100)
gMat<-matrix(sample(c(0,1),800,replace = TRUE),200,4)
pathwayCompatibilityQuartet(gMat)
}
