% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PathwayAnalysis.R
\name{pathwayFeasibility}
\alias{pathwayFeasibility}
\title{pathwayFeasibility}
\usage{
pathwayFeasibility(DAG, x)
}
\arguments{
\item{DAG}{matrix representing the DAG of restrictions.}

\item{x}{the number of mutations considered.}
}
\value{
a binary vector, which indicates feasibility or infeasibility of a set of pathways
}
\description{
pathwayFeasibility
}
\examples{
DAG<-matrix(c(4,4,4,1,2,3),3,2)
x<-4
pathwayFeasibility(DAG, x)
}
