% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PathwayAnalysis.R
\name{pathwayGenotypeCompatibility}
\alias{pathwayGenotypeCompatibility}
\title{pathwayGenotypeCompatibility}
\usage{
pathwayGenotypeCompatibility(pathway, genotype)
}
\arguments{
\item{pathway}{a vector representing the given pathway.}

\item{genotype}{a binary vector representing the given genotype.}
}
\value{
returns 1 (if the given genotype is compatible with the given pathway), and 0 otherwise
}
\description{
pathwayGenotypeCompatibility
}
\examples{
geno1 <- c(1, 0, 1, 0)
geno2 <- c(1, 1, 0, 0)
path <- c(1, 2, 3, 4)
pathwayGenotypeCompatibility(path, geno1)
pathwayGenotypeCompatibility(path, geno2)
}
