% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PathwayAnalysis.R
\name{permutations}
\alias{permutations}
\title{permutations}
\usage{
permutations(n, r, v = 1:n, set = TRUE, repeats.allowed = FALSE)
}
\arguments{
\item{n}{total number of elements in the set}

\item{r}{subset size}

\item{v}{}

\item{set}{}

\item{repeats.allowed}{}
}
\description{
permutations
}
