% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_provenance.R
\name{poset_example_data}
\alias{poset_example_data}
\title{Example .poset files}
\source{
\url{https://bsse.ethz.ch/cbg/software/ct-cbn.html}
}
\description{
Event poset used if -e is not set; if -e is set, the file is used for determining the number of events as specified in the first row
}
\details{
These files are included in the package under \verb{inst/extdata/} and can be accessed using \code{\link[=system.file]{system.file()}} or \code{\link[=get_examples]{get_examples()}}. They are unchanged from their original source.
}
