% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PathwayAnalysis.R
\name{Poset_Weighting_RCBN}
\alias{Poset_Weighting_RCBN}
\title{Poset_Weighting_RCBN}
\usage{
Poset_Weighting_RCBN(vec)
}
\arguments{
\item{vec}{The likelihood vector corresponding to a given set of posets}
}
\value{
The poset weight vector determined using the reciprocal ranking method
}
\description{
Poset_Weighting_RCBN
}
\examples{
set.seed(100)
LogLik<-runif(219)
W1<-Poset_Weighting_RCBN(LogLik)
}
