% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PathwayAnalysis.R
\name{posetWeightingRCBN}
\alias{posetWeightingRCBN}
\title{posetWeightingRCBN}
\usage{
posetWeightingRCBN(vec)
}
\arguments{
\item{vec}{The likelihood vector corresponding to a given set of posets}
}
\value{
The poset weight vector determined using the reciprocal ranking method
}
\description{
posetWeightingRCBN
}
\examples{
set.seed(100)
logLik <- runif(219)
w1 <- posetWeightingRCBN(logLik)
}
