% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rCBN-package.R
\docType{package}
\name{rCBN-package}
\alias{rCBN}
\alias{rCBN-package}
\title{rCBN: Conjunctive Bayesian Networks}
\description{
A family of R functions based on the CBN family of functions created at ETH-Zurich.
}
\author{
\strong{Maintainer}: William Choi-Kim \email{william@williamck.com}

}
\keyword{internal}
