% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_lambda}
\alias{read_lambda}
\title{Read a .lambda file}
\usage{
read_lambda(filestem)
}
\arguments{
\item{filestem}{The filename of the .lambda file without the .lambda suffix.}
}
\value{
A matrix.
}
\description{
Read a .lambda file
}
\examples{
\dontrun{
bcPath = get_examples()[1]
read_lambda(bcPath)
}
}
