% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_poset}
\alias{read_poset}
\title{Read a .poset file}
\usage{
read_poset(filestem)
}
\arguments{
\item{filestem}{The filename of the .poset file without the .poset suffix.}
}
\value{
A list containing the number of mutations and a matrix.
}
\description{
Read a .poset file
}
\examples{
bcPath = get_examples()[1]
read_poset(bcPath)
}
