% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{readLambda}
\alias{readLambda}
\title{Read a .lambda file}
\usage{
readLambda(fileStem)
}
\arguments{
\item{fileStem}{The filename of the .lambda file without the .lambda suffix.}
}
\value{
A matrix.
}
\description{
Read a .lambda file
}
\examples{
bcPath <- getExamples()[1]
readLambda(bcPath)
}
