% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{readPattern}
\alias{readPattern}
\title{Read a .pat file}
\usage{
readPattern(fileStem)
}
\arguments{
\item{fileStem}{The filename of the .pat file without the .pat suffix.}
}
\value{
A matrix.
}
\description{
Read a .pat file
}
\examples{
bcPath <- getExamples()[1]
readPattern(bcPath)
}
