% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{readTime}
\alias{readTime}
\title{Read a .time file}
\usage{
readTime(filestem)
}
\arguments{
\item{filestem}{The filename of the .time file without the .time suffix.}
}
\value{
A matrix.
}
\description{
Read a .time file
}
\examples{
bcPath <- getExamples()[1]
readPattern(bcPath)
}
