% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{visualize_cbn_model}
\alias{visualize_cbn_model}
\title{Visualize CBN Model}
\usage{
visualize_cbn_model(poset, nodeColor = "darkgreen")
}
\arguments{
\item{poset}{Poset object to visualize}

\item{nodeColor}{Color of nodes in resulting graph}
}
\value{
Plot (gg object) visualization of CBN model
}
\description{
Visualize CBN Model
}
\examples{
poset = read_poset(get_examples()[1])
visualize_cbn_model(poset)
}
