% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{visualizeCBNModel}
\alias{visualizeCBNModel}
\title{Visualize CBN Model}
\usage{
visualizeCBNModel(poset, nodeColor = "darkgreen")
}
\arguments{
\item{poset}{Poset object to visualize}

\item{nodeColor}{Color of nodes in resulting graph}
}
\value{
Plot (gg object) visualization of CBN model
}
\description{
Visualize CBN Model
}
\examples{
poset <- readPoset(getExamples()[1])
visualizeCBNModel(poset$sets)
}
