% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PathwayAnalysis.R
\name{wR}
\alias{wR}
\title{wR}
\usage{
wR(vec)
}
\arguments{
\item{vec}{The likelihood vector corresponding to a given set of posets}
}
\value{
The poset weight vector determined using the reciprocal ranking method
}
\description{
wR
}
