% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PathwayAnalysis.R
\name{WW}
\alias{WW}
\title{WW}
\usage{
WW(EdgeProb, PEmap)
}
\arguments{
\item{EdgeProb}{Marginal edge probabilities}

\item{PEmap}{Pathway-edge compatibility matrix}
}
\value{
The pathway weights (step 4 of the R-CBN algorithm)
}
\description{
WW
}
