% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PathwayAnalysis.R
\name{Base2Indexing}
\alias{Base2Indexing}
\title{Base2Indexing}
\usage{
Base2Indexing(mat)
}
\arguments{
\item{mat}{A given poset represented by a binary matrix (in B-CBN)}
}
\value{
#Poset weight vectors based on the frequency of occurence in the BCBN MCMC-sampling scheme.
}
\description{
Base2Indexing
}
