% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PathwayAnalysis.R
\name{base2IndVec}
\alias{base2IndVec}
\title{base2IndVec}
\usage{
base2IndVec(vec)
}
\arguments{
\item{vec}{a binary genotype vector}
}
\value{
a number used for indexing a given genotype
}
\description{
base2IndVec
}
\examples{
vec<-c(0,1,0,1)
base2IndVec(vec)
}
