% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CBN2Path.R
\docType{package}
\name{CBN2Path-package}
\alias{CBN2Path}
\alias{CBN2Path-package}
\title{CBN2Path: Conjunctive Bayesian Networks}
\description{
A family of R functions based on the CBN family of functions created at ETH-Zurich.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/rockwillck/CBN2Path}
  \item \url{http://dx.doi.org/10.1093/biomet/asp023}
  \item \url{http://dx.doi.org/10.1093/bioinformatics/btp505}
}

}
\author{
\strong{Maintainer}: William Choi-Kim \email{william@williamck.com}

}
\keyword{internal}
