% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctcbn.R
\name{ctcbn}
\alias{ctcbn}
\title{CT-CBN}
\usage{
ctcbn(
  datasets,
  bootstrap_mode = FALSE,
  bootstrap_samples = 0,
  random_seed = 0,
  sampling_rate = 1,
  epsilon = 2,
  num_drawn_samples = 0,
  num_em_runs = 1,
  output_stems = NULL,
  n_cores = 1
)
}
\arguments{
\item{datasets}{Vector of \code{Spock} objects with poset and pattern/lambda data or a \code{Spock} object (alias of ctcbn_single).}

\item{bootstrap_mode}{Boolean representing bootstrapping mode.}

\item{bootstrap_samples}{Number of bootstrap samples (requires \code{epsilon} > 0, \code{num_drawn_samples} = 0)}

\item{random_seed}{Random seed.}

\item{sampling_rate}{Sampling rate.}

\item{epsilon}{If between 0 and 1, the fraction of violations allowed per edge. If negative, the interval 0 to 0.5 will be sampled equidistantly with N points.}

\item{num_drawn_samples}{If > 0, the number of samples to draw from the model. If zero (default), the model will be learned from data.}

\item{num_em_runs}{Number of em runs.}

\item{output_stems}{Files to output to (defaults to tempfiles).}

\item{n_cores}{Maximum number of threads to use to parallelize.}
}
\value{
A matrix of results.
}
\description{
CT-CBN
}
\examples{
example_path = get_examples()[1]
bc = Spock$new(
  poset = read_poset(example_path)$sets,
  numMutations = read_poset(example_path)$mutations,
  genotypeMatrix = read_pattern(example_path)
)
ctcbn(bc)
ctcbn(c(bc, bc, bc))
}
