% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctcbn.R
\name{ctcbnSingle}
\alias{ctcbnSingle}
\title{CT-CBN Single Batch}
\usage{
ctcbnSingle(
  dataset,
  bootstrap_samples = 0,
  random_seed = 1,
  sampling_rate = 1,
  epsilon = 0,
  num_drawn_samples = 0,
  num_em_runs = 1
)
}
\arguments{
\item{dataset}{\code{Spock} object with poset and pattern/lambda data.}

\item{bootstrap_samples}{Number of bootstrap samples (requires \code{epsilon} > 0, \code{num_drawn_samples} = 0)}

\item{random_seed}{Random seed.}

\item{sampling_rate}{Sampling rate.}

\item{epsilon}{If between 0 and 1, the fraction of violations allowed per edge. If negative, the interval 0 to 0.5 will be sampled equidistantly with N points.}

\item{num_drawn_samples}{If > 0, the number of samples to draw from the model. If zero (default), the model will be learned from data.}

\item{num_em_runs}{Number of em runs.}
}
\value{
A list of output data.
}
\description{
CT-CBN Single Batch
}
\examples{
example_path <- getExamples()[1]
bc <- Spock$new(
    poset = readPoset(example_path)$sets,
    numMutations = readPoset(example_path)$mutations,
    genotypeMatrix = readPattern(example_path)
)
ctcbnSingle(bc)
}
