% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PathwayAnalysis.R
\name{EdgeMarginalized}
\alias{EdgeMarginalized}
\title{EdgeMarginalized}
\usage{
EdgeMarginalized(PathProb, x)
}
\arguments{
\item{PathProb}{The pathway probabilities returned in the step 3 of the R-CBN algorithm}

\item{x}{}
}
\value{
returns the marginal probability of all the potential edges
}
\description{
EdgeMarginalized
}
