% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PathwayAnalysis.R
\name{edgeMarginalized}
\alias{edgeMarginalized}
\title{edgeMarginalized}
\usage{
edgeMarginalized(PathProb, x)
}
\arguments{
\item{PathProb}{The pathway probabilities returned in the step 3 of the R-CBN algorithm}

\item{x}{The number of mutations to consider}
}
\value{
returns the marginal probability of all the potential edges
}
\description{
edgeMarginalized
}
\examples{
DAG<-matrix(c(2,2,4,1,3,3),3,2)
LAMBDA<-c(1,4,3,2.5,2)
x<-4
PathP<-pathProbCBN(DAG, LAMBDA, x)
EdgeProb<-edgeMarginalized(PathP,x)
}
