% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PathwayAnalysis.R
\name{genotypeFeasibility}
\alias{genotypeFeasibility}
\title{genotypeFeasibility}
\usage{
genotypeFeasibility(genotypes, DAG, x)
}
\arguments{
\item{genotypes}{the full set of potential binary genotypes of a given length.}

\item{DAG}{matrix representing the DAG of restrictions.}

\item{x}{the number of mutations considered.}
}
\value{
a binary vector, which indicates feasibility or infeasibility of a set of genotypes
}
\description{
genotypeFeasibility
}
\examples{
Geno4<-generateMatrixGenotypes(4)
DAG<-matrix(c(4,4,4,1,2,3),3,2)
x<-4
GenoF4<-genotypeFeasibility(Geno4,DAG,x)
}
