% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hcbn.R
\name{hcbn}
\alias{hcbn}
\title{H-CBN}
\usage{
hcbn(
  datasetObj,
  anneal = FALSE,
  temp = 0,
  annealing_steps = 0,
  outputStem = tempfile("output")
)
}
\arguments{
\item{datasetObj}{\code{Spock} object with poset and pattern/lambda data.}

\item{anneal}{If \code{TRUE}, performes a simulated annealing run starting from the poset}

\item{temp}{Temperature of simulated annealing.}

\item{annealing_steps}{Number of simulated annealing steps.}

\item{outputStem}{File to output to (defaults to a tempfile).}
}
\value{
A list of output data.
}
\description{
H-CBN
}
\examples{
example_path = get_examples()[1]
bc = Spock$new(
  poset = read_poset(example_path)$sets,
  numMutations = read_poset(example_path)$mutations,
  genotypeMatrix = read_pattern(example_path)
)
hcbn(bc)
}
