% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PathwayAnalysis.R
\name{jensenShannonDivergence}
\alias{jensenShannonDivergence}
\title{jensenShannonDivergence}
\usage{
jensenShannonDivergence(Prob1, Prob2)
}
\arguments{
\item{Prob1}{The first (discrete) probability distribution (vector)}

\item{Prob2}{The second (discrete) probability distribution (vector)}
}
\value{
Jensen Shannon Divergence between the two (discrete) probability distributions
}
\description{
jensenShannonDivergence
}
\examples{
set.seed(100)
gMat<-matrix(sample(c(0,1),12,replace = TRUE),3,4)
PathCT<-pathProbQuartetCTCBN(gMat)
PathH<-pathProbQuartetHCBN(gMat)
jensenShannonDivergence(PathCT,PathH)
}
