% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PathwayAnalysis.R
\name{PathProb_BCBN}
\alias{PathProb_BCBN}
\title{PathProb_BCBN}
\usage{
PathProb_BCBN(MAT)
}
\arguments{
\item{MAT}{transition probability matrix returned by B-CBN model}
}
\value{
vector of probabilities assigned to a set of pathways
}
\description{
PathProb_BCBN
}
