% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PathwayAnalysis.R
\name{Pathway_Feasibility}
\alias{Pathway_Feasibility}
\title{Pathway_Feasibility}
\usage{
Pathway_Feasibility(DAG, x)
}
\arguments{
\item{DAG}{matrix representing the DAG of restrictions.}

\item{x}{the number of mutations considered.}
}
\value{
a binary vector, which indicates feasibility or infeasibility of a set of pathways
}
\description{
Pathway_Feasibility
}
