% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PathwayAnalysis.R
\name{pathwayGenotypeCompatibility}
\alias{pathwayGenotypeCompatibility}
\title{pathwayGenotypeCompatibility}
\usage{
pathwayGenotypeCompatibility(Pathway, Genotype)
}
\arguments{
\item{Pathway}{a vector representing the given pathway.}

\item{Genotype}{a binary vector representing the given genotype.}
}
\value{
returns 1 (if the given genotype is compatible with the given pathway), and 0 otherwise
}
\description{
pathwayGenotypeCompatibility
}
\examples{
Geno1<-c(1,0,1,0)
Geno2<-c(1,1,0,0)
Path<-c(1,2,3,4)
pathwayGenotypeCompatibility(Path,Geno1)
pathwayGenotypeCompatibility(Path,Geno2)
}
