% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PathwayAnalysis.R
\name{Predictability}
\alias{Predictability}
\title{Predictability}
\usage{
Predictability(Prob, x)
}
\arguments{
\item{Prob}{Pathway probability vector}

\item{x}{The length of genotype vectors}
}
\value{
Predictability
}
\description{
Predictability
}
