% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{readPoset}
\alias{readPoset}
\title{Read a .poset file}
\usage{
readPoset(filestem)
}
\arguments{
\item{filestem}{The filename of the .poset file without the .poset suffix.}
}
\value{
A list containing the number of mutations and a matrix.
}
\description{
Read a .poset file
}
\examples{
bcPath <- getExamples()[1]
readPoset(bcPath)
}
