% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{visualizeCBNModel}
\alias{visualizeCBNModel}
\title{Visualize CBN Model}
\usage{
visualizeCBNModel(
  poset,
  nodeColor = "darkgreen",
  numNodes = min(4, max(poset))
)
}
\arguments{
\item{poset}{Poset object to visualize}

\item{nodeColor}{Color of nodes in resulting graph}

\item{numNodes}{Number of nodes (default is the larger number between 4 and the largest index given in the poset)}
}
\value{
Plot (gg object) visualization of CBN model
}
\description{
Visualize CBN Model
}
\examples{
poset <- readPoset(getExamples()[1])
visualizeCBNModel(poset$sets)
}
